// src/service/auth/type.ts
export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  id: number;
  username: string;
  email: string;
  firstName: string;
  lastName: string;
  gender: string;
  image: string;
}

export interface RefreshToken {
  accessToken: string;
  refreshToken: string;
}